(WIP) Moved common views under ui package.
