+++ app/src/main/java/org/hisp/dhis/android/dataentry/commons/ui/Presenter.java
@@ -1,4 +1,4 @@
-package org.hisp.dhis.android.dataentry.commons.views;
+package org.hisp.dhis.android.dataentry.commons.ui;
 
 import android.support.annotation.NonNull;
 import android.support.annotation.UiThread;
