+++ app/src/main/java/org/hisp/dhis/android/dataentry/commons/ui/View.java
@@ -1,0 +1,4 @@
+package org.hisp.dhis.android.dataentry.commons.ui;
+
+public interface View {
+}
