+++ app/src/main/java/org/hisp/dhis/android/dataentry/launcher/LauncherPresenter.java
@@ -2,7 +2,7 @@ package org.hisp.dhis.android.dataentry.launcher;
 
 import android.support.annotation.UiThread;
 
-import org.hisp.dhis.android.dataentry.commons.views.Presenter;
+import org.hisp.dhis.android.dataentry.commons.ui.Presenter;
 
 interface LauncherPresenter extends Presenter {
