+++ app/src/main/java/org/hisp/dhis/android/dataentry/reports/ReportsPresenter.java
@@ -1,6 +1,6 @@
 package org.hisp.dhis.android.dataentry.reports;
 
-import org.hisp.dhis.android.dataentry.commons.views.Presenter;
+import org.hisp.dhis.android.dataentry.commons.ui.Presenter;
 
 interface ReportsPresenter extends Presenter {
 }
