+++ app/src/main/res/layout/recyclerview_report_item.xml
@@ -18,7 +18,7 @@
         android:layout_marginRight="4dp"
         android:background="?attr/selectableItemBackground">
 
-        <org.hisp.dhis.android.dataentry.commons.views.CircleView
+        <org.hisp.dhis.android.dataentry.commons.ui.CircleView
             android:id="@+id/circleview_status_background"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
@@ -36,7 +36,7 @@
             android:contentDescription="@null"/>
     </FrameLayout>
 
-    <org.hisp.dhis.android.dataentry.commons.views.FontTextView
+    <org.hisp.dhis.android.dataentry.commons.ui.FontTextView
         android:id="@+id/textview_report_value_labels"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
